Move RandomNumberControl and ProjectData to Utilities
