+++ src/AFQMC/Hamiltonians/tests/test_hamiltonian_factory.cpp
@@ -16,7 +16,7 @@
 #include "Configuration.h"
 
 #include "OhmmsData/Libxml2Doc.h"
-#include "OhmmsApp/ProjectData.h"
+#include "ProjectData.h"
 #include "hdf/hdf_archive.h"
 
 #undef APP_ABORT
