--- src/QMCApp/tests/CMakeLists.txt
@@ -1,38 +1,0 @@
-#//////////////////////////////////////////////////////////////////////////////////////
-#// This file is distributed under the University of Illinois/NCSA Open Source License.
-#// See LICENSE file in top directory for details.
-#//
-#// Copyright (c) 2017 Jeongnim Kim and QMCPACK developers.
-#//
-#// File developed by: Mark Dewing, mdewing@anl.gov, Argonne National Laboratory
-#//
-#// File created by: Mark Dewing, mdewing@anl.gov, Argonne National Laboratory
-#//////////////////////////////////////////////////////////////////////////////////////
-
-set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${QMCPACK_UNIT_TEST_DIR})
-
-# Also includes tests for code OhmmsApp directory
-
-set(SRC_DIR qmcapp)
-set(UTEST_EXE test_${SRC_DIR})
-set(UTEST_NAME deterministic-unit_test_${SRC_DIR})
-
-add_executable(${UTEST_EXE} test_project_data.cpp test_rng_control.cpp)
-
-target_link_libraries(${UTEST_EXE} catch_main qmc)
-if(USE_OBJECT_TARGET)
-  target_link_libraries(
-    ${UTEST_EXE}
-    qmcestimators
-    qmcdriver
-    qmcham
-    qmcwfs
-    qmcparticle
-    qmcutil
-    platform_omptarget_LA)
-  if(BUILD_AFQMC)
-    target_link_libraries(${UTEST_EXE} afqmc)
-  endif(BUILD_AFQMC)
-endif()
-
-add_unit_test(${UTEST_NAME} 1 1 "${QMCPACK_UNIT_TEST_DIR}/${UTEST_EXE}")
