+++ src/Utilities/tests/test_rng_control.cpp
@@ -16,7 +16,7 @@
 //#include "Utilities/RandomGenerator.h"
 #include "Message/Communicate.h"
 #include "OhmmsData/Libxml2Doc.h"
-#include "OhmmsApp/RandomNumberControl.h"
+#include "RandomNumberControl.h"
 
 
 #include <stdio.h>
